DS project components/requirements.txt
streamlit
langchain
langchain_groq
langchain_community
pandas
matplotlib
numpy
seaborn
python_dotenv
ipykernel
langchain_ollama